fix logo not applying system theme
